Changed the temp file removal so that if the temp file no longer exists, a warning message is not printed, since the goal is delete it anyway.

+++ dApp.py
@@ -58,7 +58,8 @@ class TempFileHolder(object):
 		try:
 			import os
 			for f in self._tempFiles:
-				os.remove(f)
+				if not os.path.exists(f):
+					continue
 				try:
 					os.remove(f)
 				except OSError, e:
